check read-only status for derived methods (IDEADEV-38128)

+++ inspections/impl/com/intellij/codeInspection/unneededThrows/RedundantThrows.java
@@ -1,6 +1,7 @@
 package com.intellij.codeInspection.unneededThrows;
 
 import com.intellij.analysis.AnalysisScope;
+import com.intellij.codeInsight.CodeInsightUtilBase;
 import com.intellij.codeInsight.ExceptionUtil;
 import com.intellij.codeInsight.daemon.GroupNames;
 import com.intellij.codeInspection.*;
@@ -233,6 +234,9 @@ public class RedundantThrows extends GlobalJavaInspectionTool {
             }
           }
         }
+
+        //check read-only status for derived methods
+        if (!CodeInsightUtilBase.preparePsiElementsForWrite(refsToDelete)) return;
 
         for (final PsiJavaCodeReferenceElement aRefsToDelete : refsToDelete) {
           aRefsToDelete.delete();
